claim button and functionality

+++ src/components/requests/Request.js
@@ -1,28 +1,31 @@
 import { Link, Navigate, useNavigate } from "react-router-dom"
 
-export const Request = ({ requestObject , currentUser , employees , getAllRequests , requestNumber}) => {
-
+export const Request = ({ requestObject , currentUser , employees , getAllRequests , requestNumber, value }) => {
     let assignedEmployee = null
     let navigate = useNavigate()
 
-    if (requestObject.employeeRequests.length > 0) {
+    if (requestObject?.employeeRequests?.length > 0) {
         const requestEmployeeRelationship = requestObject.employeeRequests[0]
         assignedEmployee = employees.find(employee => employee.id === requestEmployeeRelationship.employeeId)
     }
 
-    const userEmployeeId = employees.find(employee => employee.userId === currentUser.id)
+    const userEmployee = employees.find(employee => employee.userId === currentUser.id)
 
-    // const canClose = () => {
-    //     if (currentUser.staff){
-    //     if (userEmployeeId?.id === assignedEmployee?.id && requestObject.dateCompleted === "") {
-    //         return <button className="ticket__finish"
-    //                 onClick={closeTicket}
-    //                 >Close Request</button>
-    //     } else {
-    //         return ""
-    //             }        
-    //     }
-    // }
+    function refreshPage() {
+        window.location.reload(false)
+    }
+    
+    const canClose = () => {
+        if (currentUser.staff){
+        if (userEmployee?.id === assignedEmployee?.id && requestObject.dateCompleted === "") {
+            return <button className="ticket__finish"
+                    onClick={closeTicket}
+                    >Close</button>
+        } else {
+            return ""
+                }        
+        }
+    }
     
     const deleteButton = () => {
         if (!currentUser.staff && !requestObject.isComplete) {
@@ -33,9 +36,10 @@ export const Request = ({ requestObject , currentUser , employees , getAllReques
                     .then(() => {
                         getAllRequests()
                     })
+                    .then(refreshPage)
             }} className="ticket__delete">Cancel Request</button> 
         } 
-        else if (currentUser.staff && !requestObject.isComplete) {
+        else if (currentUser.staff && !requestObject.isComplete && requestObject?.employeeRequests.length === 0) {
             return <button onClick={() => {
                 fetch(`http://localhost:8088/serviceRequests/${requestObject.id}`, {
                     method: "DELETE",
@@ -43,89 +47,164 @@ export const Request = ({ requestObject , currentUser , employees , getAllReques
                 .then(() => {
                     getAllRequests()
                 })
-            }} className="ticket__delete">Deny Request</button>
+                .then(refreshPage)
+            }} className="ticket__delete">Deny</button>
         }
         else {
-            return <button onClick={() => {
-            }} className="ticket__delete">See Details</button>
+            return ""
+            // <button onClick={() => {
+            // }} className="ticket__delete">Details</button>""
         }
     }
 
     const editButton = () => {
         if (!currentUser.staff && !requestObject.isComplete) {
-            return <Link to={`/request/${requestObject.id}/edit`} state={{ user: {currentUser}}}>Edit
+            return <button><Link to={`/request/${requestObject.id}/edit`} state={{ user: {currentUser}}}>Edit
                     </Link>
+                    </button>
             }
             else {
                 return ""
             }   
         }
 
-    // const closeTicket = () => {
-    //     const copy = {
-    //         userId: requestObject.id,
-    //         description: requestObject.description,
-    //         emergency: requestObject.emergency,
-    //         dateCompleted: new Date()
-    //     }
+    
+    const closeTicket = () => {
+        const copy = {
+                customerId: requestObject.customerId,
+                description: requestObject.description,
+                specialtyId: requestObject.specialtyId,
+                dateCompleted: new Date(),
+                isComplete: true,
+                invoiceId: requestObject.invoiceId
+            }
+    
+            fetch(`http://localhost:8088/serviceRequests/${requestObject.id}`, {
+                method: "PUT",
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify(copy)
+    
+            })
+                .then(response => response.json)
+                .then(getAllRequests)
+                .then(refreshPage)
+        }
+    
+    const submitReviewButton = () => {
+        return <button><Link to={`/request/${requestObject.id}/submitreview`} state={`{ userId: ${currentUser.id}}`}>Submit Review
+                    </Link>
+                    </button>
+    }
+    const claimButton = () => {
+        if (currentUser.staff && requestObject?.employeeRequests?.length === 0) {
+            return <button
+                            onClick={() => {
+                                fetch(`http://localhost:8088/employeeRequests`, {
+                                    method: "POST",
+                                    headers: {
+                                        "Content-Type": "application/json"
+                                    },
+                                    body: JSON.stringify({
+                                        employeeId: userEmployee.id,
+                                        serviceRequestId: requestObject.id
+                                    })
+                                })
+                                    .then(response => response.json())
+                                    .then(() => {
+                                        getAllRequests()
+                                    })
+                                    .then(refreshPage)
+                            }}
+                            >Claim Request</button>
+        } else if (requestObject?.employeeRequests?.length > 0 && requestObject.isComplete === false){
+            return `STATUS: IN PROCESS. Assigned to ${assignedEmployee !== null ? assignedEmployee?.user?.firstName : ""}`
+        } else {
+            return `STATUS: COMPLETE. Completed by ${assignedEmployee !== null ? assignedEmployee?.user?.firstName : ""}`
+        }
+    }
 
-    //     fetch(`http://localhost:8088/serviceTickets/${requestObject.id}`, {
-    //         method: "PUT",
-    //         headers: {
-    //             "Content-Type": "application/json"
-    //         },
-    //         body: JSON.stringify(copy)
+    const display = () => {
+        if (currentUser.staff) {
+            if (value === true) {
+            return <>
+                    <section key={`key--${requestObject.id}`} className="ticket">
+                        <header>
+                            <Link to={`/request/${requestObject.id}`} state={{requestNumber: {requestNumber}}}>Request {requestNumber}
+                        </Link>
+                        </header>
+                            <section>{requestObject?.description}</section>
+                        <footer>
+                            {
+                                claimButton()
+                            }
+                            {
+                                editButton()
+                            }
+                            {
+                                canClose()
+                            }
+                            {
+                                deleteButton()
+                            }
+                        </footer>
+                    </section>
+        </>
+        } else {
+            return <>
+                <section key={`key--${requestObject?.serviceRequest?.id}`} className="tickets">
+                    <header>Request {requestNumber}</header>
+                        <section>{requestObject?.serviceRequest?.description}</section>
+                            <footer>
+                            {
+                                editButton()
+                            }
+                            {/* {
+                                canClose()
+                            } */}
+                            {
+                                deleteButton()
+                            }
+                            </footer>
+                </section>
+            </>
+        }
+    }
+        else {
+            return <>
+                    <section key={`key--${requestObject.id}`} className="ticket">
+                        <header>
+                            <Link to={`/request/${requestObject.id}`} state={{requestNumber: {requestNumber}}}>Request {requestNumber}
+                        </Link>
+                        </header>
+                            <section>{requestObject?.description}</section>
+                        <footer>
+                            {
+                                claimButton()
+                            }
+                            {
+                                editButton()
+                            }
+                            {
+                                canClose()
+                            }
+                            {
+                                deleteButton()
+                            }
+                            {
+                                submitReviewButton()
+                            }
+                        </footer>
+                    </section>
+        </>
+        }
+    }
 
-    //     })
-    //         .then(response => response.json)
-    //         .then(getAllRequests)
-    // }
-
-    // const buttonOrNoButton = () => {
-    //     if (currentUser.staff) {
-    //         return <button
-    //                         onClick={() => {
-    //                             fetch(`http://localhost:8088/employeeTickets`, {
-    //                                 method: "POST",
-    //                                 headers: {
-    //                                     "Content-Type": "application/json"
-    //                                 },
-    //                                 body: JSON.stringify({
-    //                                     employeeId: userEmployeeId.id,
-    //                                     serviceTicketId: requestObject.id
-    //                                 })
-    //                             })
-    //                                 .then(response => response.json())
-    //                                 .then(() => {
-    //                                     getAllRequests()
-    //                                 })
-    //                         }}
-    //                         >Claim</button>
-    //     } else {
-    //         return ""
-    //     }
-    // }
-
-    return <section key={`key--${requestObject.id}`} className="ticket">
-            <header>Request {requestNumber}</header>
-                <section>{requestObject.description}</section>
-            <footer>
+    return  <>
                 {
-                    requestObject?.employeeRequests.length
-                        ? `IN PROCESS: assigned to ${assignedEmployee !== null ? assignedEmployee?.user?.firstName : ""}`
-                        :  ""
-                        // buttonOrNoButton()
-                }
-                {
-                    editButton()
-                }
-                {/* {
-                    canClose()
-                } */}
-                {
-                    deleteButton()
-                }
-            </footer>
-
-        </section>
+                    display()
+                } 
+            </>
 }
+
